# server/internal/platform/httpx/openapi.yaml
openapi: 3.1.0
info:
  title: Ship API
  version: 0.1.0
  description: The contract exposed by the Ship control-plane API.
servers:
  - url: /
paths:
...
  /auth/password:
    post:
      operationId: changePassword
      x-ship-permission: authenticated
      security: [{ sessionCookie: [] }]
      summary: Change the current user's password and rotate their session
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: Password changed and replacement session issued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: CSRF validation failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ship_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32 }
...
  schemas:
...
    Role:
      type: string
      enum: [owner, admin, deployer, viewer]
    User:
      type: object
      required: [id, email, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
...
    Session:
      type: object
      required: [user, csrfToken, expiresAt]
      properties:
        user: { $ref: "#/components/schemas/User" }
        csrfToken: { type: string }
        expiresAt: { type: string, format: date-time }
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, maxLength: 128 }
        newPassword: { type: string, minLength: 12, maxLength: 128 }
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"